+++ packages/orq-rc/RELEASES.md
@@ -2659,3 +2659,13 @@ Based on:
 - [typescript v3.10.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.24 - packages/orq-rc
+
+## 2025-07-29 10:31:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.592.1 (2.667.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.25 - packages/orq-rc